+++ Cargo.toml
@@ -1,41 +1,62 @@
 [package]
 name = "sybot_lib"
 authors = [ "Samuel Nösslböck <samuel.noesslboeck@gmail.com>" ]
-version = "0.7.7"
+version = "0.7.8"
 description = "A simple library to control groups of components and robots"
 edition = "2021"
 repository = "https://github.com/SamuelNoesslboeck/stepper_lib"
 license-file = "LICENSE"
-keywords = [ "framework", "robotics" ]
+keywords = [ "framework", "robotics", "automation" ]
 
 [lib]
 name = "sybot_lib"
 path = "src/lib.rs"
 
-[[bin]]
+[[example]]
 name = "intpr"
 path = "bin/interpreter.rs"
 
-[[bin]]
+[[example]]
+name = "mqtt"
+path = "examples/mqtt.rs"
+required-features = [ "mqtt" ]
+
+[[example]]
 name = "webserver"
 path = "bin/webserver.rs"
+required-features = [ "http" ]
 
 [dependencies]
-actix = "0.13.0"
-actix-web = "4"
-actix-web-actors = "4.1.0"
+actix = { version = "0.13.0", optional = true } 
+actix-web = { version = "4", optional = true }
+actix-web-actors = { version = "4.1.0", optional = true } 
 
 colored = "2.0.0"
 colored_json = "3.0.1"
 
 gcode = "0.6.1"
 glam = "0.21.3"
-paho-mqtt = "0.12.0"
+
+paho-mqtt = { version = "0.12.0", optional = true }
+
 serde = { version = "1.0.104", features = ["derive"] }
 serde_json = "1.0.48"
 
-stepper_lib = { path = "../stepper_lib", features = [ "std" ] } # { git = "https://github.com/SamuelNoesslboeck/stepper_lib" }
+stepper_lib = { version = "0.11.1" } # { git = "https://github.com/SamuelNoesslboeck/stepper_lib" }
 
 [features]
-default = ["dbg-funcs"]
+default = [ "dbg-funcs" ]
 dbg-funcs = []
+
+# platform features
+rasp = [ "stepper_lib/rasp" ]
+
+# remote features
+http = [ "dep:actix", "dep:actix-web", "dep:actix-web-actors" ]
+mqtt = [ "dep:paho-mqtt" ]
+
+[profile.dev]
+opt-level = 0
+
+[profile.release]
+opt-level = 3
